generator/pyproject.toml
[tool.poetry]
name = "generator"
version = "0.1.0"
description = ""
authors = ["statefb <tmgf.15.su@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.18.2"
google-cloud-storage = "^1.27.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
